Prefer dist-info metadata to egg-info when parsing wheel version
In the event that a wheel has both a .dist-info folder and a .egg-info
file, the wheel_version() function currently picks whichever file was
the first one returned from os.listdir() in the path where the wheel was
extracted. Since FileMetadata.get_metadata returns a KeyError for
anything other than PKG-INFO, if the .egg-info file happens to be first
then check_compatibility will raise an UnsupportedWheel exception.